clinic/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
fastapi==0.115.0
uvicorn==0.30.6
openpyxl==3.1.2
networkx==3.2.1
Werkzeug==3.0.1
pydantic==2.9.2
python-multipart==0.0.6
python-dotenv==1.0.1
requests==2.32.3
matplotlib==3.9.2
Pillow==10.4.0
google-api-python-client==2.99.0
google-generativeai==0.3.2
graphviz==0.21
pandas==2.2.3
opencv-python==4.7.0.72
numpy==1.26.4
